Made the worker more efficient, and added prep vars for Quest Levels

+++ src/me/quaz3l/qQuests/API/QuestModels/Builders/BuildonSomething.java
@@ -7,9 +7,9 @@ import me.quaz3l.qQuests.API.QuestModels.onSomething;
 
 public class BuildonSomething {
 	private String message = "";
-	private Integer money = 0;
-	private Integer health = 0;
-	private Integer hunger = 0;
+	private double money = 0;
+	private int health = 0;
+	private int hunger = 0;
 	private HashMap<Integer, ArrayList<Integer>> items = new HashMap<Integer, ArrayList<Integer>>();
 	private HashMap<Integer, String> permissionsAdd = new HashMap<Integer, String>();
 	private HashMap<Integer, String> permissionsTake = new HashMap<Integer, String>();
@@ -22,15 +22,15 @@ public class BuildonSomething {
 		this.message = s;
 		return this;
 	}
-	public BuildonSomething money(Integer i) {
+	public BuildonSomething money(double i) {
 		this.money = i;
 		return this;
 	}
-	public BuildonSomething health(Integer i) {
+	public BuildonSomething health(int i) {
 		this.health = i;
 		return this;
 	}
-	public BuildonSomething hunger(Integer i) {
+	public BuildonSomething hunger(int i) {
 		this.hunger = i;
 		return this;
 	}
@@ -38,13 +38,13 @@ public class BuildonSomething {
 	public String message() {
 		return this.message;
 	}
-	public Integer money() {
+	public double money() {
 		return this.money;
 	}
-	public Integer health() {
+	public int health() {
 		return this.health;
 	}
-	public Integer hunger() {
+	public int hunger() {
 		return this.hunger;
 	}
 	public HashMap<Integer, ArrayList<Integer>> items() {
